代码修改。

+++ src/doc/数据库模型/scripts/lottery数据库建表.sql
@@ -1,7 +1,7 @@
 create table lottery_ticket
 (
      lottery_run_time date,
-     lottery_period_num number(7,0),
+     lottery_period_num int(7),
      lottery_red_first varchar(4),
      lottery_red_second varchar(4),
      lottery_red_third varchar(4),
@@ -9,7 +9,7 @@ create table lottery_ticket
      lottery_red_fifth varchar(4),
      lottery_red_sixth varchar(4),
      lottery_blue_first varchar(4),
-     lottery_sale_volume number(10),
+     lottery_sale_volume int(10),
      primary key (lottery_period_num)
 );
